asset/ignition: move content under bootstrap
The content package only contains bootstrap assets. It's more clear that
that is the case if that package is under the bootstrap package, rather
than being adjacent to it.
