+++ pkg/asset/ignition/bootstrap/content/bootkube.go
