+++ pkg/asset/ignition/bootstrap/content/doc.go
@@ -1,3 +1,3 @@
 // Package content contains the contents of files and systemd units to be added
-// to Ignition configs.
+// to bootstrap Ignition configs.
 package content
