+++ pkg/asset/ignition/bootstrap/content/kubelet.go
