+++ pkg/asset/ignition/bootstrap/content/tectonic.go
